fix another use-after-free crash in downloader
if disconnect is called while the async http request was still ongoing, then the "this" pointer for onDownloadComplete will have been freed.
a shared_ptr will make sure that the object stays alive until the callback referencing the DownloadManager object is done.
this might cause hanging if the async request takes a long time, for some reason, but idk, better than crashing (for now)