bump version to 40.01
i have finally realized there is a bug/inconsistency with the SDL X11 mouse handler.
when calling SDL_WarpMouseInWindow (in Mouse.cpp through env->setMousePos(center)),
there's like a 25% chance that one of the SDL_EVENT_MOUSE_MOTION events in the next
SDL_PumpEvents call we make will contain this warp.
sometimes it's the first motion event in the queue, sometimes it's the second, sometimes it's never.

we have ZERO way to distinguish this event from the other, non-synthetic motion events,
so this results in the ingame cursor making a huge teleport from the previous mouse centering.
this doesn't happen at all with the windows or wayland SDL_VIDEO_DRIVERs.

tl;dr: more time well spent on mouse