2 minor fixes
in my windows vm, the 16x AA rendertarget for the main menu button wasn't working (because it's not supported by the driver).
just use 4x AA, which is basically guaranteed. we can check for supported features later, but 16x was way overkill anyways

also the check for \\?\ paths on windows was broken, oops