soloud: run it on a separate thread, and add init timeout
there is basically no downside to this.
the mixing function itself (which runs every time the audio playback thread requests data) can be quite heavy, so it's good to offload that.
this allows us to call functions (like init) with timeouts, so we don't hang if some backend initialization hangs (like SDL+SDL_AUDIO_DRIVER=jack does, for me).